pyproject.toml
[project]
name = "z-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-aiplatform>=1.70.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-vertexai>=0.1.0",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.2.0",
    "langgraph-cli[inmem]>=0.1.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.1",
    "websockets>=12.0",
]
